Removed Drawing Component
